Adding new workflow for instalation, on new device
This commit adds script for automatic instalation, this is usualy
used on debian like systems, if you encounter any problem with the
script please use a Issue for this to be fixed.

Adding plugins on zsh, missing on old version, used not to be in
the repo, so on `ohmyzsh` I added them.

This commit also adds Updated readme on how to install, caution on
the script, it also download all my dependencies for a new system
using apt.

Fixes issue on vimrc-config where I was using multiple vimrc
template, so now I'm only using one. The other wasn't loading
correcly.

Tested on:
  - Linux penguin 4.19.79-07511-ge32b3719f26b #1 x86_64 GNU/Linux